backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-CORS==4.0.1
Flask-Limiter==3.8.0
SQLAlchemy==2.0.36
Werkzeug==3.1.3
python-dotenv==1.0.1
requests==2.32.3
anthropic==0.40.0
gunicorn==23.0.0
pytest==8.3.4
pytest-flask==1.3.0
